Remove c-style casts

+++ src/game/magic/RuneDraw.cpp
@@ -155,8 +155,8 @@ void ARX_SPELLS_UpdateSymbolDraw() {
 				light->pos += Vec3f(0.f, -120.f, 0.f);
 				
 				float rr = Random::getf();
-				light->fallstart=140.f+(float)io->flarecount*0.333333f+rr*5.f;
-				light->fallend=220.f+(float)io->flarecount*0.5f+rr*5.f;
+				light->fallstart = 140.f + io->flarecount * 0.333333f + rr * 5.f;
+				light->fallend = 220.f + io->flarecount * 0.5f + rr * 5.f;
 				light->intensity=1.6f;
 				light->rgb.r=0.01f*io->flarecount*2;
 				light->rgb.g=0.009f*io->flarecount*2;
